Weird fix to access @prisma/client module
I don't know why this was such an issue, well, I kinda do.
In order to use import.meta.url to get the working directory,
the typescript module setting can't be "commonjs" which would
solve this problem. So I need to set the module type to NodeNext
and the moduleResolution to NodeNext in the tsconfig to properly see
and access @prisma/client. Of course that creates a new problem, because
you can't use import statements without file extensions in NodeNext for
whatever reason, so I had to also set "allowImportingTsExtensions" to true,
but that can't be set without also setting noEmit to true as well.
I don't even know what noEmit does, but everything is working again...